chore: Move migration to script
