+++ server/api/utils.js
@@ -3,107 +3,19 @@ import subDays from "date-fns/sub_days";
 import debug from "debug";
 import Router from "koa-router";
 import { AuthenticationError } from "../errors";
-import {
-  Document,
-  Team,
-  User,
-  AuthenticationProvider,
-  UserAuthentication,
-  Attachment,
-} from "../models";
+import { Document, Attachment } from "../models";
 import { Op, sequelize } from "../sequelize";
 import parseAttachmentIds from "../utils/parseAttachmentIds";
 
 const router = new Router();
 const log = debug("utils");
 
-function authMiddleware(ctx, next: () => Promise<*>) {
-  const { token } = ctx.body;
+router.post("utils.gc", async (ctx) => {
+  const { token, limit = 500 } = ctx.body;
 
   if (process.env.UTILS_SECRET !== token) {
     throw new AuthenticationError("Invalid secret token");
   }
-
-  return next();
-}
-
-router.post("utils.migrateAuthentication", authMiddleware, async (ctx) => {
-  let { page = 0, limit = 1000 } = ctx.body;
-  const cache = {};
-
-  const process = async (page: number) => {
-    log(`Migrating authentication data… page ${page}`);
-
-    const users = await User.findAll({
-      limit,
-      offset: page * limit,
-      paranoid: false,
-      where: {
-        serviceId: {
-          [Op.ne]: "email",
-        },
-      },
-      include: [
-        {
-          model: Team,
-          as: "team",
-          required: true,
-          paranoid: false,
-        },
-      ],
-    });
-
-    for (const user of users) {
-      for (const provider of ["slack", "google"]) {
-        const providerId = user.team[`${provider}Id`];
-        if (!providerId) {
-          continue;
-        }
-
-        let authenticationProviderId = cache[providerId];
-        if (!authenticationProviderId) {
-          const [
-            authenticationProvider,
-          ] = await AuthenticationProvider.findOrCreate({
-            where: {
-              name: provider,
-              providerId,
-              teamId: user.teamId,
-            },
-          });
-
-          cache[providerId] = authenticationProviderId =
-            authenticationProvider.id;
-        }
-
-        try {
-          await UserAuthentication.create({
-            authenticationProviderId,
-            providerId: user.serviceId,
-            teamId: user.teamId,
-            userId: user.id,
-          });
-        } catch (err) {
-          console.log(
-            `serviceId ${user.serviceId} exists, for user ${user.id}`
-          );
-          continue;
-        }
-      }
-    }
-
-    return users.length === limit ? process(page + 1) : undefined;
-  };
-
-  await process(page);
-
-  ctx.body = {
-    success: true,
-  };
-});
-
-router.post("utils.gc", authMiddleware, async (ctx) => {
-  const { limit = 500 } = ctx.body;
 
   log(`Permanently destroying upto ${limit} documents older than 30 days…`);
