+++ server/scripts/bootstrap.js
@@ -1,0 +1,4 @@
+// @flow
+require("dotenv").config({ silent: true });
+
+process.env.SINGLE_RUN = true;
